airflow/requirements.txt
apache-airflow-providers-google >= 6.2.0
google-api-core >= 1.31.4
pyarrow >= 5.0.0
pandas >= 1.3.5
kaggle == 1.5.12
openpyxl >= 3.0.10
Flask >= 1.1.2
SQLAlchemy==1.4.29
croniter >= 1.0.15
google-cloud-bigquery >= 2.31.0
google-cloud-bigquery-datatransfer >= 3.4.1
google-cloud-bigquery-storage >= 2.10.1
airflow-provider-great-expectations >= 0.2.2
great_expectations == 0.16.13
sqlalchemy-bigquery==1.6.1
